Remove old files/add notes for potential updates
Darcie and Jeremy use case documents are deleted as that information
is now contained in the RAD.  Notes are underneath almost all use
cases for differences that could/should be corrected/improved.

Need to look out for more changes and once revisions are done then
it should all be reviewed again for consistency between use cases.