Update the activities to reference an external ip
Adds the system_vars.xml file to the values directory.
This holds the external address of the server for use creating
network request urls.  Has the string resource server_address
with the initial part of the server address.  Everyone will need
to create this file locally.  The .gitignore is set to not track
this file, so there will be no more merge conflicts.