Matching works in the android portion of the app
The match screen displays basic info on the user, and allows them to
respond with block, study, or pass.  Each response sends a database
request to keep track of this information.  Block adds them to the
user's block list, so they do not appear again in matches.  Study
takes the user to send a new message.  Pass just loads the next user.

Next version will rank the matches sent to the application based on
in-common classes and in-common responses.